+++ compiler-core/src/ast.rs
@@ -1365,11 +1365,6 @@ pub enum BitArrayOption<Value> {
         location: SrcSpan,
     },
 
-    // TODO: remove deprecated syntax
-    Binary {
-        location: SrcSpan,
-    },
-
     Int {
         location: SrcSpan,
     },
@@ -1379,11 +1374,6 @@ pub enum BitArrayOption<Value> {
     },
 
     Bits {
-        location: SrcSpan,
-    },
-
-    // TODO: remove deprecated syntax
-    BitString {
         location: SrcSpan,
     },
 
@@ -1454,8 +1444,6 @@ impl<A> BitArrayOption<A> {
     pub fn location(&self) -> SrcSpan {
         match self {
             BitArrayOption::Bytes { location }
-            | BitArrayOption::Binary { location }
-            | BitArrayOption::BitString { location }
             | BitArrayOption::Int { location }
             | BitArrayOption::Float { location }
             | BitArrayOption::Bits { location }
@@ -1477,12 +1465,10 @@ impl<A> BitArrayOption<A> {
 
     pub fn label(&self) -> EcoString {
         match self {
-            BitArrayOption::Binary { .. } => "bytes".into(),
             BitArrayOption::Bytes { .. } => "bytes".into(),
             BitArrayOption::Int { .. } => "int".into(),
             BitArrayOption::Float { .. } => "float".into(),
             BitArrayOption::Bits { .. } => "bits".into(),
-            BitArrayOption::BitString { .. } => "bits".into(),
             BitArrayOption::Utf8 { .. } => "utf8".into(),
             BitArrayOption::Utf16 { .. } => "utf16".into(),
             BitArrayOption::Utf32 { .. } => "utf32".into(),
@@ -1497,24 +1483,6 @@ impl<A> BitArrayOption<A> {
             BitArrayOption::Size { .. } => "size".into(),
             BitArrayOption::Unit { .. } => "unit".into(),
         }
-    }
-
-    /// Returns `true` if the bit array option is [`Binary`].
-    ///
-    /// [`Binary`]: BitArrayOption::Binary
-    #[must_use]
-    pub fn is_binary(&self) -> bool {
-        matches!(self, Self::Binary { .. })
-    }
-
-    /// Returns `true` if the bit array option is [`BitString`].
-    ///
-    /// The deprecated `bit_string` variable specifically!
-    ///
-    /// [`BitString`]: BitArrayOption::BitString
-    #[must_use]
-    pub fn is_bit_string(&self) -> bool {
-        matches!(self, Self::BitString { .. })
     }
 }