+++ compiler-core/src/javascript/pattern.rs
@@ -399,12 +399,26 @@ impl<'module_ctx, 'expression_gen, 'a> Generator<'module_ctx, 'expression_gen, '
             Pattern::StringPrefix {
                 left_side_string,
                 right_side_assignment,
+                left_side_assignment,
                 ..
             } => {
                 self.push_string_prefix_check(subject.clone(), left_side_string);
-                self.push_string_prefix_slice(utf16_no_escape_len(left_side_string));
                 if let AssignName::Variable(right) = right_side_assignment {
+                    self.push_string_prefix_slice(utf16_no_escape_len(left_side_string));
                     self.push_assignment(subject.clone(), right);
+                    // We remove the string slicing that was needed to push the correct assignment
+                    // this way the following assignments will only be sliced if necessary.
+                    self.pop();
+                }
+                if let Some((left, _)) = left_side_assignment {
+                    // "foo" as prefix <> rest
+                    //       ^^^^^^^^^ In case the left prefix of the pattern matching is given an
+                    //                 alias we bind it to a local variable so that it can be
+                    //                 correctly referenced inside the case branch.
+                    // let prefix = "foo";
+                    // ^^^^^^^^^^^^^^^^^^^ we're adding this assignment inside the if clause
+                    //                     the case branch gets translated into.
+                    self.push_assignment(super::expression::string(left_side_string), left);
                 }
                 self.pop();
                 Ok(())
@@ -497,7 +511,7 @@ impl<'module_ctx, 'expression_gen, 'a> Generator<'module_ctx, 'expression_gen, '
                             Ok(())
                         }
 
-                        [Opt::Bytes { .. } | Opt::Binary { .. }] => {
+                        [Opt::Bytes { .. }] => {
                             self.push_rest_from(offset.bytes);
                             self.traverse_pattern(subject, &segment.value)?;
                             self.pop();
@@ -505,29 +519,27 @@ impl<'module_ctx, 'expression_gen, 'a> Generator<'module_ctx, 'expression_gen, '
                             Ok(())
                         }
 
-                        [Opt::Bytes { .. } | Opt::Binary { .. }, Opt::Size { value: size, .. }]
-                        | [Opt::Size { value: size, .. }, Opt::Bytes { .. } | Opt::Binary { .. }] => {
-                            match &**size {
-                                Pattern::Int { value, .. } => {
-                                    let start = offset.bytes;
-                                    let increment = value.parse::<usize>().expect(
-                                        "part of an Int node should always parse as integer",
-                                    );
-                                    offset.increment(increment);
-                                    let end = offset.bytes;
+                        [Opt::Bytes { .. }, Opt::Size { value: size, .. }]
+                        | [Opt::Size { value: size, .. }, Opt::Bytes { .. }] => match &**size {
+                            Pattern::Int { value, .. } => {
+                                let start = offset.bytes;
+                                let increment = value
+                                    .parse::<usize>()
+                                    .expect("part of an Int node should always parse as integer");
+                                offset.increment(increment);
+                                let end = offset.bytes;
 
-                                    self.push_binary_from_slice(start, end);
-                                    self.traverse_pattern(subject, &segment.value)?;
-                                    self.pop();
-                                    Ok(())
-                                }
-
-                                _ => Err(Error::Unsupported {
-                                    feature: "This bit array size option in patterns".into(),
-                                    location: segment.location,
-                                }),
+                                self.push_binary_from_slice(start, end);
+                                self.traverse_pattern(subject, &segment.value)?;
+                                self.pop();
+                                Ok(())
                             }
-                        }
+
+                            _ => Err(Error::Unsupported {
+                                feature: "This bit array size option in patterns".into(),
+                                location: segment.location,
+                            }),
+                        },
 
                         _ => Err(Error::Unsupported {
                             feature: "This bit array segment option in patterns".into(),