+++ compiler-core/src/type_/tests.rs
@@ -516,7 +516,6 @@ fn infer_module_type_retention_test() {
     assert_eq!(
         module.type_info,
         ModuleInterface {
-            type_only_unqualified_imports: Vec::new(),
             origin: Origin::Src,
             package: "thepackage".into(),
             name: "ok".into(),
@@ -782,7 +781,7 @@ fn bit_array3() {
 
 #[test]
 fn bit_array4() {
-    assert_infer!("let <<x:binary>> = <<1>> x", "BitArray");
+    assert_infer!("let <<x:bytes>> = <<1>> x", "BitArray");
 }
 
 #[test]
@@ -824,7 +823,7 @@ fn bit_array10() {
 #[test]
 fn bit_array11() {
     assert_infer!(
-        "let a = <<1>> let <<x:bytes>> = <<1, a:2-bits>> x",
+        "let a = <<1>> let <<x:bits>> = <<1, a:2-bits>> x",
         "BitArray"
     );
 }